# src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: openapi@inss.ch
  description: |-
    Forked and modified by International Network Support & Service - Glas.
    This document contains some more endpoints containing APIs that don't have only common formats like application/json and other commonly used formats, and contains also float numbers and Umlaute.
    This is to test & check if a tool can handle also those formats.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://inss.ch
  title: Swagger Petstore - OpenAPI 3.0
  version: 1.0.0
...
servers:
- url: /api/v3
...
paths:
...
  /store/inventory:
    get:
      description: Returns a map of status codes to quantities
      operationId: getInventory
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  format: int32
                  type: integer
                type: object
          description: successful operation
      security:
      - api_key: []
      summary: Returns pet inventories by status
      tags:
      - store
      x-accepts: application/json
      x-tags:
      - tag: store
  /store/order:
    post:
      description: Place a new order in the store
      operationId: placeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: successful operation
        "405":
          description: Invalid input
      summary: Place an order for a pet
      tags:
      - store
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: store
  /store/order/{orderId}:
    delete:
      description: For valid response try integer IDs with value < 1000. Anything
        above 1000 or nonintegers will generate API errors
      operationId: deleteOrder
      parameters:
      - description: ID of the order that needs to be deleted
        explode: false
        in: path
        name: orderId
        required: true
        schema:
          example: 1001
          format: int64
          type: integer
        style: simple
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      summary: Delete purchase order by ID
      tags:
      - store
      x-accepts: application/json
      x-tags:
      - tag: store
    get:
      description: For valid response try integer IDs with value <= 5 or > 10. Other
        values will generated exceptions
      operationId: getOrderById
      parameters:
      - description: ID of order that needs to be fetched
        explode: false
        in: path
        name: orderId
        required: true
        schema:
          example: 1001
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      summary: Find purchase order by ID
      tags:
      - store
      x-accepts: application/json
      x-tags:
      - tag: store
      - tag: troublemaker
...
components:
...
  schemas:
...
    Order:
      example:
        petId: 198772
        rabatt: 25.9
        quantity: 7
        pricebrutto: 345.45
        pricenetto: 300.0
        id: 10
        shipDate: 2017-07-21T17:32:28Z
        complete: true
        status: approved
      properties:
        id:
          description: Id of the order.
          example: 10
          format: int64
          type: integer
        petId:
          description: PetId
          example: 198772
          format: int64
          type: integer
        pricebrutto:
          description: Brutto price
          example: 345.45
          format: float
          type: number
        pricenetto:
          description: Netto price
          example: 300.0
          format: float
          type: number
        rabatt:
          description: Rabatt reduction.
          example: 25.9
          format: float
          type: number
        quantity:
          description: Quantity
          example: 7
          format: int32
          type: integer
        shipDate:
          description: Shipping date.
          example: 2017-07-21T17:32:28Z
          format: date-time
          type: string
        status:
          description: Order Status
          enum:
          - placed
          - approved
          - delivered
          example: approved
          type: string
        complete:
          description: Order completed.
          example: true
          type: boolean
      type: object
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey